Revert "feat(web): add project home overview"
This reverts commit c334c27871fcd0fd23e3b1011227e86ddbdc44b4.
