+++ apps/web/src/domains/projects/project-sections.ts
@@ -9,7 +9,6 @@ import {
   EyeOffIcon,
   Fingerprint,
   FlaskConical,
-  HouseIcon,
   ImportIcon,
   Key,
   type LucideIcon,
@@ -45,20 +44,12 @@ interface ProjectSection {
 
 const PROJECT_SECTIONS: readonly ProjectSection[] = [
   {
-    key: "home",
-    label: "Home",
-    icon: HouseIcon,
-    group: "observe",
-    path: (slug) => `/projects/${slug}`,
-    isActive: (pathname, slug) => pathname === `/projects/${slug}` || pathname === `/projects/${slug}/`,
-  },
-  {
     key: "sessions",
     label: "Sessions",
     icon: MessagesSquareIcon,
     group: "observe",
-    path: (slug) => `/projects/${slug}/sessions`,
-    isActive: (pathname, slug) => pathname.startsWith(`/projects/${slug}/sessions`),
+    path: (slug) => `/projects/${slug}`,
+    isActive: (pathname, slug) => pathname === `/projects/${slug}` || pathname === `/projects/${slug}/`,
   },
   {
     key: "users",
