+++ apps/web/src/routes/_authenticated/projects/$projectSlug/cost/-components/use-go-to-model-sessions.ts
@@ -6,7 +6,7 @@ export function useGoToModelSessions(projectSlug: string) {
   const navigate = useNavigate()
   return (model: string) =>
     void navigate({
-      to: "/projects/$projectSlug/sessions",
+      to: "/projects/$projectSlug",
       params: { projectSlug },
       search: { tab: "sessions", filters: serializeFilters({ models: [{ op: "in", value: [model] }] }) },
     })
